moved divisions.vue into [level].vue
